[fix] Run scripts only from the manager dialog (FIB-340)
The Scripts menu listed every runnable script and ran it on click. That was
fine while scripts were inline and instant. It is not fine now one can drive
the microscope on a worker thread: a menu item has nowhere to show that a
script is still going, and no Stop button, so a ten-minute stage survey
started from the menu could not be interrupted at all.

The menu is now a way in, not a way to run -- Manage scripts… and Open scripts
folder, nothing else. That also settles what it should list: nothing. Names
alone cannot express the folder, a load error, or that a script writes or needs
hardware, and the dialog shows all of it right beside the button that starts it.

Built once instead of on aboutToShow, since nothing in it depends on the folder
any more. Opening the Tools menu no longer imports every script in it.

Also fixes the microscope example in SCRIPTING.md, which was wrong in two ways
and had never been executed: lamella.state.stage_position does not exist (it is
lamella.stage_position), and microscope.acquire_image() ignores save=True --
fibsem.acquire.acquire_image() is the one that writes the file. The example is
now extracted from the markdown and run against a demo microscope, and it
writes the images it claims to.

Runner tests move out of test_script_menu.py into test_script_runner.py. They
were only there because the menu used to be the thing that ran scripts.
